refactor: rename popup/Checkbox -> popup/CheckboxForPopup
